feat(speaking): normalize AI score to 0-10 average scale across UI and API

+++ src/main/java/com/datn/engflow/model/dto/response/SpeakingSubmissionResponse.java
@@ -21,7 +21,7 @@ public record SpeakingSubmissionResponse(
         Integer scoreGrammar,
         Integer scoreVocabulary,
         Integer scoreFluency,
-        Integer scoreTotal,
+        Double scoreTotal,
         Double pronunciationAccuracy,
         Double pronunciationFluency,
         Double pronunciationCompleteness,
